# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Órdenes de Compra a Proveedores
  description: >
    Contrato de referencia para la implementación del backend. Este archivo es
    el contrato preciso contra el cual se implementa y se mantiene a mano.
  version: "2026-09-24"
  contact:
    name: Henry Molina
    url: https://henry-molina.dev/
servers:
  - url: http://localhost:8080/api
    description: Entorno local (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /sucursales:
    get:
      tags: [Sucursales]
      summary: Listar sucursales
      operationId: listarSucursales
      description: >
        Catálogo de referencia (puntos de venta y su formato) que alimenta los
        selectores de "sucursal destino" al crear una orden y de "sucursal" al
        dar de alta un usuario GERENTE_SUCURSAL. Cualquier usuario autenticado
        puede consultarlo: no expone información sensible y lo necesitan tanto
        pantallas de Compras como de Administración.
      responses:
        "200":
          description: Listado de sucursales
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Sucursal" }
    post:
      tags: [Sucursales]
      summary: Crear sucursal
      operationId: crearSucursal
      description: >
        Da de alta un punto de venta con su formato, que decide qué categorías
        de producto se le pueden pedir y si sus órdenes exigen cliente (Venta
        Directa). Reservado a ADMIN, como el resto de la administración de
        catálogos.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SucursalRequest" }
      responses:
        "201":
          description: Sucursal creada
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sucursal" }
        "400":
          $ref: "#/components/responses/InvalidRequestError"
        "403":
          description: El usuario no es ADMIN
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /sucursales/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    put:
      tags: [Sucursales]
      summary: Actualizar sucursal
      operationId: actualizarSucursal
      description: >
        Edita el nombre o el formato de una sucursal, o la activa/desactiva
        mediante el campo activo (una sucursal desactivada no puede elegirse
        como destino de órdenes nuevas; las órdenes que ya la tienen como
        destino siguen siendo editables). El formato no se puede cambiar una
        vez que la sucursal tiene órdenes: 409 con código formato_bloqueado.
        Para "cambiarlo" se desactiva la sucursal y se crea otra. Reservado a
        ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SucursalRequest" }
      responses:
        "200":
          description: Sucursal actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sucursal" }
        "400":
          $ref: "#/components/responses/InvalidRequestError"
        "403":
          description: El usuario no es ADMIN
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: El formato no se puede cambiar porque la sucursal ya tiene órdenes (código formato_bloqueado)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /proveedores:
    get:
      tags: [Proveedores]
      summary: Listar proveedores
      operationId: listarProveedores
      description: >
        Catálogo de proveedores disponibles para armar órdenes de compra y si
        están activos. Disponible para ADMIN y COMPRADOR (el comprador debe
        poder elegir proveedor al crear una orden). La clave de webhook solo se
        devuelve a ADMIN; para COMPRADOR el campo webhook_api_key viene nulo,
        porque quien la conoce puede registrar eventos en nombre del proveedor.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Listado de proveedores
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Proveedor" }
    post:
      tags: [Proveedores]
      summary: Crear proveedor
      operationId: crearProveedor
      description: >
        Da de alta un nuevo proveedor en el catálogo y le genera automáticamente
        una clave de webhook para que pueda reportar eventos logísticos más
        adelante. Reservado a ADMIN, como el resto de la administración de
        catálogos: alta de proveedores es una decisión de negocio, no una
        operación del día a día de Compras.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProveedorRequest" }
      responses:
        "201":
          description: Proveedor creado
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Proveedor" }
        "400":
          $ref: "#/components/responses/InvalidRequestError"
  /proveedores/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    put:
      tags: [Proveedores]
      summary: Actualizar proveedor
      operationId: actualizarProveedor
      description: >
        Edita los datos de contacto de un proveedor existente, o lo
        activa/desactiva mediante el campo activo (un proveedor desactivado
        deja de estar disponible para nuevas órdenes). Reservado a ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProveedorRequest" }
      responses:
        "200":
          description: Proveedor actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Proveedor" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /proveedores/{id}/regenerar-clave-webhook:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    patch:
      tags: [Proveedores]
      summary: Regenerar la clave de webhook del proveedor
      operationId: regenerarClaveWebhookProveedor
      description: >
        Genera una clave de webhook nueva y devuelve el proveedor con ella. La
        clave anterior deja de autenticar de inmediato, asi que la nueva debe
        hacerse llegar al proveedor por un canal aparte. Sirve ante una fuga o
        para rotarla periodicamente. Reservado a ADMIN.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Clave regenerada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Proveedor" }
        "403":
          description: El usuario no es ADMIN
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /productos:
    get:
      tags: [Productos]
      summary: Listar productos
      operationId: listarProductos
      description: >
        Catálogo de productos que un COMPRADOR recorre para armar el detalle de
        una orden (con filtro opcional por categoría). Abierto también a ADMIN
        para las mismas pantallas de administración de catálogos.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: categoria_id
          in: query
          schema: { type: integer }
          description: Filtrar por categoría
        - name: sucursal_id
          in: query
          schema: { type: integer }
          description: >
            Devuelve solo los productos cuya categoría está permitida para el
            formato de esa sucursal, es decir, los que se pueden pedir para ella.
      responses:
        "200":
          description: Listado de productos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Producto" }
    post:
      tags: [Productos]
      summary: Crear producto
      operationId: crearProducto
      description: >
        Da de alta un nuevo producto en el catálogo, incluida la categoría que
        determina en qué formatos de sucursal se puede pedir. Reservado a ADMIN:
        un COMPRADOR elige productos existentes al armar una orden, no define el
        catálogo.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductoRequest" }
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Producto" }
  /productos/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    put:
      tags: [Productos]
      summary: Actualizar producto
      operationId: actualizarProducto
      description: >
        Edita los datos de un producto existente (precio, unidades, categoría),
        o lo activa/desactiva mediante el campo activo. Reservado a ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductoRequest" }
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Producto" }
  /clientes:
    get:
      tags: [Clientes]
      summary: Listar clientes
      operationId: listarClientes
      description: >
        Catálogo de clientes finales, obligatorio para elegir uno al crear una
        orden con sucursal destino de Venta Directa. Abierto a COMPRADOR (quien
        arma esas órdenes) y a ADMIN (administración de catálogos).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Listado de clientes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Cliente" }
    post:
      tags: [Clientes]
      summary: Crear cliente
      operationId: crearCliente
      description: >
        Da de alta un nuevo cliente final en el catálogo. Reservado a ADMIN,
        como el resto del mantenimiento de catálogos.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClienteRequest" }
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cliente" }
  /clientes/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    put:
      tags: [Clientes]
      summary: Actualizar cliente
      operationId: actualizarCliente
      description: >
        Edita los datos de un cliente existente, o lo activa/desactiva mediante
        el campo activo. Reservado a ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClienteRequest" }
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cliente" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT propio emitido por /auth/login (sin OAuth2)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Credencial por proveedor, usada exclusivamente por el endpoint de webhooks
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Clave generada por el cliente para evitar reprocesar la misma operación ante reintentos
...
  responses:
    InvalidRequestError:
      description: Error de validación de entrada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    FormatoSucursal:
      type: string
      enum: [FERRETERIA, FERRETERIA_CONSTRUCCION, VENTA_DIRECTA]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [type, code, message]
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, state_conflict_error, api_error]
            code: { type: string, example: categoria_no_permitida }
            message: { type: string }
            param: { type: string, nullable: true }
    ListEnvelope:
      type: object
      properties:
        object: { type: string, enum: [list] }
        has_more: { type: boolean }
    Sucursal:
      type: object
      properties:
        object: { type: string, enum: [sucursal] }
        id: { type: integer }
        nombre: { type: string }
        formato: { $ref: "#/components/schemas/FormatoSucursal" }
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Proveedor:
      type: object
      properties:
        object: { type: string, enum: [proveedor] }
        id: { type: integer }
        nombre: { type: string }
        email: { type: string, format: email }
        telefono: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        webhook_api_key:
          type: string
          nullable: true
          description: >
            Generada automaticamente al crear el proveedor. Se entrega al proveedor real fuera de
            banda. ADMIN puede consultarla via GET y rotarla con
            PATCH /proveedores/{id}/regenerar-clave-webhook.
        descuento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 99.99
          description: >
            Descuento maximo (en porcentaje) que se puede negociar sobre el precio de catalogo de
            cualquier producto de este proveedor. 0 junto con aumento_maximo_pct = 0 significa que
            no se negocia (se compra al precio de catalogo).
        aumento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Aumento maximo (en porcentaje) que se acepta sobre el precio de catalogo.
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProveedorRequest:
      type: object
      required: [nombre, email]
      properties:
        nombre: { type: string, minLength: 1, maxLength: 150 }
        email: { type: string, format: email }
        telefono: { type: string, maxLength: 30 }
        direccion: { type: string, maxLength: 250 }
        activo:
          type: boolean
          description: Omitido en creacion equivale a true; omitido en edicion preserva el valor actual.
        descuento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 99.99
          description: Porcentaje maximo de descuento negociable. Omitido en creacion equivale a 0; omitido en edicion preserva el valor actual. Maximo dos decimales.
        aumento_maximo_pct:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Porcentaje maximo de aumento aceptable. Omitido en creacion equivale a 0; omitido en edicion preserva el valor actual. Maximo dos decimales.
    SucursalRequest:
      type: object
      required: [nombre, formato]
      properties:
        nombre: { type: string, minLength: 1, maxLength: 150 }
        formato: { $ref: "#/components/schemas/FormatoSucursal" }
        activo:
          type: boolean
          description: Omitido en creacion equivale a true; omitido en edicion preserva el valor actual.
    Producto:
      type: object
      properties:
        object: { type: string, enum: [producto] }
        id: { type: integer }
        codigo: { type: string }
        nombre: { type: string }
        categoria_id: { type: integer }
        precio: { type: number, format: double }
        unidad_venta: { type: string }
        unidad_compra: { type: string }
        factor_conversion: { type: number, format: double }
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProductoRequest:
      type: object
      required: [codigo, nombre, categoria_id, precio, unidad_venta, unidad_compra]
      properties:
        codigo: { type: string, maxLength: 30 }
        nombre: { type: string, minLength: 1, maxLength: 150 }
        categoria_id: { type: integer }
        precio: { type: number, format: double, exclusiveMinimum: true, minimum: 0 }
        unidad_venta: { type: string }
        unidad_compra: { type: string }
        factor_conversion: { type: number, format: double, default: 1 }
        activo:
          type: boolean
          description: Omitido en creacion equivale a true; omitido en edicion preserva el valor actual.
    Cliente:
      type: object
      properties:
        object: { type: string, enum: [cliente] }
        id: { type: integer }
        nombre: { type: string }
        tipo: { type: string, enum: [INDUSTRIAL, CONTRATISTA, COMERCIO] }
        email: { type: string, format: email, nullable: true }
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ClienteRequest:
      type: object
      required: [nombre, tipo]
      properties:
        nombre: { type: string, minLength: 1, maxLength: 150 }
        tipo: { type: string, enum: [INDUSTRIAL, CONTRATISTA, COMERCIO] }
        email: { type: string, format: email }
        activo:
          type: boolean
          description: Omitido en creacion equivale a true; omitido en edicion preserva el valor actual.
...